requirements.txt
requests==2.31.0
PySocks==1.7.1
pydantic==1.10.9
moviepy==1.0.3
pycryptodomex==3.18.0
instagrapi
customtkinter==5.1.3
Pillow
ffmpeg-python